refactor: rename file to built-in-provider.ts
